Move while and until loop parser
